requirements.txt
# Django: The web framework for building web applications
Django>=5.1.1,<5.2

# Django REST Framework: Toolkit for building Web APIs
djangorestframework>=3.15.2,<3.16

# psycopg: PostgreSQL adapter for Python
psycopg>=3.2.1,<3.3

# drf-spectacular: OpenAPI 3 schema generation for Django REST Framework
drf-spectacular>=0.27.2,<0.28
